fix(util): use rustls as the curl TLS backend
Making this change allows cargo to be used on Windows 7 and 8.1 against
the current Fastly TLS configuration.

This does the minimal amount of plumbing required for curl to be built
with rustls as the TLS backend, and for rustls to be configured with the
system store's root certificate bundle. (Using webpki_roots would be
inappropriate, since enterprise users may have custom CA roots
configured.)

I've made no effort to make this configurable thus far. Presumably we'd
only want to do this on Windows, if at all.

Technically, this would fix #8154.